Remove duplicated code from database wrapper #PAMP-45 Remove duplicated code from database wrapper
