+++ pamagent/hooks/sqlite_hook.py
@@ -4,7 +4,7 @@ from pamagent.hooks.dbapi2 import (CursorWrapper as DBAPI2CursorWrapper, Connect
                                    ConnectionFactory as DBAPI2ConnectionFactory)
 from pamagent.trace import register_database_client, DatabaseTrace
 from pamagent.transaction_cache import current_transaction
-from pamagent.wrapper import FuncWrapper, callable_name, wrap_object
+from pamagent.wrapper import wrap_object
 
 
 DEFAULT = object()
@@ -33,23 +33,11 @@ class CursorWrapper(DBAPI2CursorWrapper):
 class ConnectionWrapper(DBAPI2ConnectionWrapper):
     __cursor_wrapper__ = CursorWrapper
 
-    def __enter__(self):
-        transaction = current_transaction()
-        name = callable_name(self.__wrapped__.__enter__)
-        with FuncWrapper(transaction, name):
-            self.__wrapped__.__enter__()
-        return self
-
-    def __exit__(self, exc, value, tb, *args, **kwargs):
-        transaction = current_transaction()
-        name = callable_name(self.__wrapped__.__exit__)
-        with FuncWrapper(transaction, name):
-            if exc is None and value is None and tb is None:
-                with DatabaseTrace(transaction, 'COMMIT', self._pam_dbapi2_module, self._pam_connect_params):
-                    return self.__wrapped__.__exit__(exc, value, tb)
-            else:
-                with DatabaseTrace(transaction, 'ROLLBACK', self._pam_dbapi2_module, self._pam_connect_params):
-                    return self.__wrapped__.__exit__(exc, value, tb)
+    @staticmethod
+    def is_commit_on_exit(exc, value, tb):
+        if exc is None and value is None and tb is None:
+            return True
+        return False
 
     def execute(self, sql, parameters=DEFAULT):
         transaction = current_transaction()
